Extract performance evidence contracts
Move performance evidence DTOs into a dedicated contract module while preserving the app.contracts.performance_workspace compatibility exports.

Evidence:

- python -m ruff format --check touched files: passed

- python -m ruff check touched files: passed

- python -m mypy touched source files: passed

- python -m pytest tests/unit/test_performance_workspace_evidence.py tests/unit/test_performance_workspace_capabilities.py tests/unit/test_performance_workspace_response.py: 18 passed

- make check: 1,059 unit/contract tests passed

- make ci: 207 integration tests and 1,266 coverage tests passed, total coverage 94.05%, pip-audit no known vulnerabilities with governed PYSEC-2026-161 exception
