+++ src/app/services/performance_workspace_response.py
@@ -6,12 +6,12 @@ from typing import Any, Protocol, TypeAlias
 
 from app.contracts.performance_attribution import AttributionSummaryView
 from app.contracts.performance_contribution import ContributionSummaryView
+from app.contracts.performance_evidence import PerformanceEvidenceView
 from app.contracts.performance_workspace import (
     MoneyWeightedReturnSummary,
     PerformanceBenchmarkOptionView,
     PerformanceChartPoint,
     PerformanceComparativeSummary,
-    PerformanceEvidenceView,
     PerformanceWorkspaceCapabilities,
     PerformanceWorkspaceResponse,
 )
